[codegen] Promote codegen lib2 to lib
This means deleting all of our previous codegen code, and aceepting that
we are entering a new & wonderous world.

In addition this merges parse2_plan into the main codegen_plan, and
replaces the gpos & layout input files with gpos2 & layout2.
