+++ resources/codegen_inputs/layout.rs
@@ -1,8 +1,10 @@
+// path (from compile crate) to the generated parse module for this table.
+#![parse_module(font_tables::layout)]
+
 /// [Script List Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#script-list-table-and-script-record)
-#[offset_host]
-ScriptList<'a> {
+table ScriptList {
     /// Number of ScriptRecords
-    #[compute_count(script_records)]
+    #[compile(array_len($script_records))]
     script_count: BigEndian<u16>,
     /// Array of ScriptRecords, listed alphabetically by script tag
     #[count(script_count)]
@@ -10,7 +12,7 @@ ScriptList<'a> {
 }
 
 /// [Script Record](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#script-list-table-and-script-record)
-ScriptRecord {
+record ScriptRecord {
     /// 4-byte script tag identifier
     script_tag: BigEndian<Tag>,
     /// Offset to Script table, from beginning of ScriptList
@@ -18,22 +20,21 @@ ScriptRecord {
 }
 
 /// [Script Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#script-table-and-language-system-record)
-#[offset_host]
-Script<'a> {
+table Script {
     /// Offset to default LangSys table, from beginning of Script table
     /// — may be NULL
     #[nullable]
     default_lang_sys_offset: BigEndian<Offset16<LangSys>>,
     /// Number of LangSysRecords for this script — excluding the
     /// default LangSys
-    #[compute_count(lang_sys_records)]
+    #[compile(array_len($lang_sys_records))]
     lang_sys_count: BigEndian<u16>,
     /// Array of LangSysRecords, listed alphabetically by LangSys tag
     #[count(lang_sys_count)]
     lang_sys_records: [LangSysRecord],
 }
 
-LangSysRecord {
+record LangSysRecord {
     /// 4-byte LangSysTag identifier
     lang_sys_tag: BigEndian<Tag>,
     /// Offset to LangSys table, from beginning of Script table
@@ -41,18 +42,17 @@ LangSysRecord {
 }
 
 /// [Language System Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#language-system-table)
-LangSys<'a> {
+table LangSys {
     /// = NULL (reserved for an offset to a reordering table)
-    #[hidden]
-    #[compile_type(u16)]
-    #[compute(0)]
-    lookup_order_offset: BigEndian<Offset16>,
+    #[skip_getter]
+    #[compile(0)]
+    lookup_order_offset: BigEndian<u16>,
     /// Index of a feature required for this language system; if no
     /// required features = 0xFFFF
     required_feature_index: BigEndian<u16>,
     /// Number of feature index values for this language system —
     /// excludes the required feature
-    #[compute_count(feature_indices)]
+    #[compile(array_len($feature_indices))]
     feature_index_count: BigEndian<u16>,
     /// Array of indices into the FeatureList, in arbitrary order
     #[count(feature_index_count)]
@@ -60,38 +60,34 @@ LangSys<'a> {
 }
 
 /// [Feature List Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#feature-list-table)
-#[offset_host]
-FeatureList<'a> {
+table FeatureList {
     /// Number of FeatureRecords in this table
-    #[compute_count(feature_records)]
+    #[compile(array_len($feature_records))]
     feature_count: BigEndian<u16>,
     /// Array of FeatureRecords — zero-based (first feature has
     /// FeatureIndex = 0), listed alphabetically by feature tag
     #[count(feature_count)]
-    #[to_owned(self.feature_records_to_owned())]
     feature_records: [FeatureRecord],
 }
 
 /// Part of [FeatureList]
-#[skip_to_owned]
-FeatureRecord {
+record FeatureRecord {
     /// 4-byte feature identification tag
     feature_tag: BigEndian<Tag>,
     /// Offset to Feature table, from beginning of FeatureList
+    #[read_offset_with($feature_tag)]
     feature_offset: BigEndian<Offset16<Feature>>,
 }
 
 /// [Feature Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#feature-table)
-
-#[skip_to_owned]
-#[offset_host]
-Feature<'a> {
+#[read_args(feature_tag: Tag)]
+table Feature {
     /// Offset from start of Feature table to FeatureParams table, if defined for the feature and present, else NULL
     #[nullable]
-    #[skip_offset_getter]
+    #[read_offset_with($feature_tag)]
     feature_params_offset: BigEndian<Offset16<FeatureParams>>,
     /// Number of LookupList indices for this feature
-    #[compute_count(lookup_list_indices)]
+    #[compile(array_len($lookup_list_indices))]
     lookup_index_count: BigEndian<u16>,
     /// Array of indices into the LookupList — zero-based (first
     /// lookup is LookupListIndex = 0)
@@ -100,11 +96,10 @@ Feature<'a> {
 }
 
 /// [Lookup List Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#lookup-list-table)
-#[offset_host]
-#[no_compile]
-LookupList<'a> {
+#[skip_compile]
+table LookupList {
     /// Number of lookups in this table
-    #[compute_count(lookup_offsets)]
+    #[compile(array_len($lookup_offsets))]
     lookup_count: BigEndian<u16>,
     /// Array of offsets to Lookup tables, from beginning of LookupList
     /// — zero based (first lookup is Lookup index = 0)
@@ -113,15 +108,14 @@ LookupList<'a> {
 }
 
 /// [Lookup Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#lookup-table)
-#[offset_host]
-#[no_compile]
-Lookup<'a> {
+#[skip_compile]
+table Lookup {
     /// Different enumerations for GSUB and GPOS
     lookup_type: BigEndian<u16>,
     /// Lookup qualifiers
     lookup_flag: BigEndian<u16>,
     /// Number of subtables for this lookup
-    #[compute_count(subtable_offsets)]
+    #[compile(array_len($subtable_offsets))]
     sub_table_count: BigEndian<u16>,
     /// Array of offsets to lookup subtables, from beginning of Lookup
     /// table
@@ -134,12 +128,12 @@ Lookup<'a> {
 }
 
 /// [Coverage Format 1](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#coverage-format-1)
-CoverageFormat1<'a> {
+table CoverageFormat1 {
     /// Format identifier — format = 1
-    #[compute(1)]
+    #[format = 1]
     coverage_format: BigEndian<u16>,
     /// Number of glyphs in the glyph array
-    #[compute_count(glyph_array)]
+    #[compile(array_len($glyph_array))]
     glyph_count: BigEndian<u16>,
     /// Array of glyph IDs — in numerical order
     #[count(glyph_count)]
@@ -147,12 +141,12 @@ CoverageFormat1<'a> {
 }
 
 /// [Coverage Format 2](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#coverage-format-2)
-CoverageFormat2<'a> {
+table CoverageFormat2 {
     /// Format identifier — format = 2
-    #[compute(2)]
+    #[format = 2]
     coverage_format: BigEndian<u16>,
     /// Number of RangeRecords
-    #[compute_count(range_records)]
+    #[compile(array_len($range_records))]
     range_count: BigEndian<u16>,
     /// Array of glyph ranges — ordered by startGlyphID.
     #[count(range_count)]
@@ -160,7 +154,7 @@ CoverageFormat2<'a> {
 }
 
 /// Used in [CoverageFormat2]
-RangeRecord {
+record RangeRecord {
     /// First glyph ID in the range
     start_glyph_id: BigEndian<u16>,
     /// Last glyph ID in the range
@@ -170,23 +164,20 @@ RangeRecord {
 }
 
 /// [Coverage Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#coverage-table)
-#[format(u16)]
-enum CoverageTable<'a> {
-    #[version(1)]
-    Format1(CoverageFormat1<'a>),
-    #[version(2)]
-    Format2(CoverageFormat2<'a>),
+format u16 CoverageTable {
+    Format1(CoverageFormat1),
+    Format2(CoverageFormat2),
 }
 
 /// [Class Definition Table Format 1](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#class-definition-table-format-1)
-ClassDefFormat1<'a> {
+table ClassDefFormat1 {
     /// Format identifier — format = 1
-    #[compute(1)]
+    #[format = 1]
     class_format: BigEndian<u16>,
     /// First glyph ID of the classValueArray
     start_glyph_id: BigEndian<u16>,
     /// Size of the classValueArray
-    #[compute_count(class_value_array)]
+    #[compile(array_len($class_value_array))]
     glyph_count: BigEndian<u16>,
     /// Array of Class Values — one per glyph ID
     #[count(glyph_count)]
@@ -194,12 +185,12 @@ ClassDefFormat1<'a> {
 }
 
 /// [Class Definition Table Format 2](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#class-definition-table-format-2)
-ClassDefFormat2<'a> {
+table ClassDefFormat2 {
     /// Format identifier — format = 2
-    #[compute(2)]
+    #[format = 2]
     class_format: BigEndian<u16>,
     /// Number of ClassRangeRecords
-    #[compute_count(class_range_records)]
+    #[compile(array_len($class_range_records))]
     class_range_count: BigEndian<u16>,
     /// Array of ClassRangeRecords — ordered by startGlyphID
     #[count(class_range_count)]
@@ -207,7 +198,8 @@ ClassDefFormat2<'a> {
 }
 
 /// Used in [ClassDefFormat2]
-ClassRangeRecord {
+#[validation_method(validate_glyph_range)]
+record ClassRangeRecord {
     /// First glyph ID in the range
     start_glyph_id: BigEndian<u16>,
     /// Last glyph ID in the range
@@ -216,16 +208,14 @@ ClassRangeRecord {
     class: BigEndian<u16>,
 }
 
-#[format(u16)]
-enum ClassDef<'a> {
-    #[version(1)]
-    Format1(ClassDefFormat1<'a>),
-    #[version(2)]
-    Format2(ClassDefFormat2<'a>),
+/// A [Class Definition Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#class-definition-table)
+format u16 ClassDef {
+    Format1(ClassDefFormat1),
+    Format2(ClassDefFormat2),
 }
 
 /// [Sequence Lookup Record](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#sequence-lookup-record)
-SequenceLookupRecord {
+record SequenceLookupRecord {
     /// Index (zero-based) into the input glyph sequence
     sequence_index: BigEndian<u16>,
     /// Index (zero-based) into the LookupList
@@ -233,16 +223,15 @@ SequenceLookupRecord {
 }
 
 /// [Sequence Context Format 1](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#sequence-context-format-1-simple-glyph-contexts)
-#[offset_host]
-SequenceContextFormat1<'a> {
+table SequenceContextFormat1 {
     /// Format identifier: format = 1
-    #[compute(1)]
+    #[format = 1]
     format: BigEndian<u16>,
     /// Offset to Coverage table, from beginning of
     /// SequenceContextFormat1 table
     coverage_offset: BigEndian<Offset16<CoverageTable>>,
     /// Number of SequenceRuleSet tables
-    #[compute_count(seq_rule_set_offsets)]
+    #[compile(array_len($seq_rule_set_offsets))]
     seq_rule_set_count: BigEndian<u16>,
     /// Array of offsets to SequenceRuleSet tables, from beginning of
     /// SequenceContextFormat1 table (offsets may be NULL)
@@ -252,10 +241,9 @@ SequenceContextFormat1<'a> {
 }
 
 /// Part of [SequenceContextFormat1]
-#[offset_host]
-SequenceRuleSet<'a> {
+table SequenceRuleSet {
     /// Number of SequenceRule tables
-    #[compute_count(seq_rule_offsets)]
+    #[compile(array_len($seq_rule_offsets))]
     seq_rule_count: BigEndian<u16>,
     /// Array of offsets to SequenceRule tables, from beginning of the
     /// SequenceRuleSet table
@@ -264,15 +252,15 @@ SequenceRuleSet<'a> {
 }
 
 /// Part of [SequenceContextFormat1]
-SequenceRule<'a> {
+table SequenceRule {
     /// Number of glyphs in the input glyph sequence
-    #[compute(plus_one(self.input_sequence.len()))]
+    #[compile(plus_one($input_sequence.len()))]
     glyph_count: BigEndian<u16>,
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Array of input glyph IDs—starting with the second glyph
-    #[count_with(minus_one, glyph_count)]
+    #[count_expr(minus_one($glyph_count))]
     input_sequence: [BigEndian<u16>],
     /// Array of Sequence lookup records
     #[count(seq_lookup_count)]
@@ -280,10 +268,9 @@ SequenceRule<'a> {
 }
 
 /// [Sequence Context Format 2](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#sequence-context-format-2-class-based-glyph-contexts)
-#[offset_host]
-SequenceContextFormat2<'a> {
+table SequenceContextFormat2 {
     /// Format identifier: format = 2
-    #[compute(2)]
+    #[format = 2]
     format: BigEndian<u16>,
     /// Offset to Coverage table, from beginning of
     /// SequenceContextFormat2 table
@@ -292,7 +279,7 @@ SequenceContextFormat2<'a> {
     /// SequenceContextFormat2 table
     class_def_offset: BigEndian<Offset16<ClassDef>>,
     /// Number of ClassSequenceRuleSet tables
-    #[compute_count(class_seq_rule_set_offsets)]
+    #[compile(array_len($class_seq_rule_set_offsets))]
     class_seq_rule_set_count: BigEndian<u16>,
     /// Array of offsets to ClassSequenceRuleSet tables, from beginning
     /// of SequenceContextFormat2 table (may be NULL)
@@ -302,10 +289,9 @@ SequenceContextFormat2<'a> {
 }
 
 /// Part of [SequenceContextFormat2]
-#[offset_host]
-ClassSequenceRuleSet<'a> {
+table ClassSequenceRuleSet {
     /// Number of ClassSequenceRule tables
-    #[compute_count(class_seq_rule_offsets)]
+    #[compile(array_len($class_seq_rule_offsets))]
     class_seq_rule_count: BigEndian<u16>,
     /// Array of offsets to ClassSequenceRule tables, from beginning of
     /// ClassSequenceRuleSet table
@@ -314,16 +300,16 @@ ClassSequenceRuleSet<'a> {
 }
 
 /// Part of [SequenceContextFormat2]
-ClassSequenceRule<'a> {
+table ClassSequenceRule {
     /// Number of glyphs to be matched
-    #[compute(plus_one(self.input_sequence.len()))]
+    #[compile(plus_one($input_sequence.len()))]
     glyph_count: BigEndian<u16>,
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Sequence of classes to be matched to the input glyph sequence,
     /// beginning with the second glyph position
-    #[count_with(minus_one, glyph_count)]
+    #[count_expr(minus_one($glyph_count))]
     input_sequence: [BigEndian<u16>],
     /// Array of SequenceLookupRecords
     #[count(seq_lookup_count)]
@@ -331,16 +317,15 @@ ClassSequenceRule<'a> {
 }
 
 /// [Sequence Context Format 3](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#sequence-context-format-3-coverage-based-glyph-contexts)
-#[offset_host]
-SequenceContextFormat3<'a> {
+table SequenceContextFormat3 {
     /// Format identifier: format = 3
-    #[compute(3)]
+    #[format = 3]
     format: BigEndian<u16>,
     /// Number of glyphs in the input sequence
-    #[compute_count(coverage_offsets)]
+    #[compile(array_len($coverage_offsets))]
     glyph_count: BigEndian<u16>,
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Array of offsets to Coverage tables, from beginning of
     /// SequenceContextFormat3 subtable
@@ -351,28 +336,22 @@ SequenceContextFormat3<'a> {
     seq_lookup_records: [SequenceLookupRecord],
 }
 
-#[format(u16)]
-#[offset_host]
-enum SequenceContext<'a> {
-    #[version(1)]
-    Format1(SequenceContextFormat1<'a>),
-    #[version(2)]
-    Format2(SequenceContextFormat2<'a>),
-    #[version(3)]
-    Format3(SequenceContextFormat3<'a>),
+format u16 SequenceContext {
+    Format1(SequenceContextFormat1),
+    Format2(SequenceContextFormat2),
+    Format3(SequenceContextFormat3),
 }
 
 /// [Chained Sequence Context Format 1](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#chained-sequence-context-format-1-simple-glyph-contexts)
-#[offset_host]
-ChainedSequenceContextFormat1<'a> {
+table ChainedSequenceContextFormat1 {
     /// Format identifier: format = 1
-    #[compute(1)]
+    #[format = 1]
     format: BigEndian<u16>,
     /// Offset to Coverage table, from beginning of
     /// ChainSequenceContextFormat1 table
     coverage_offset: BigEndian<Offset16<CoverageTable>>,
     /// Number of ChainedSequenceRuleSet tables
-    #[compute_count(chained_seq_rule_set_offsets)]
+    #[compile(array_len($chained_seq_rule_set_offsets))]
     chained_seq_rule_set_count: BigEndian<u16>,
     /// Array of offsets to ChainedSeqRuleSet tables, from beginning of
     /// ChainedSequenceContextFormat1 table (may be NULL)
@@ -382,10 +361,9 @@ ChainedSequenceContextFormat1<'a> {
 }
 
 /// Part of [ChainedSequenceContextFormat1]
-#[offset_host]
-ChainedSequenceRuleSet<'a> {
+table ChainedSequenceRuleSet {
     /// Number of ChainedSequenceRule tables
-    #[compute_count(chained_seq_rule_offsets)]
+    #[compile(array_len($chained_seq_rule_offsets))]
     chained_seq_rule_count: BigEndian<u16>,
     /// Array of offsets to ChainedSequenceRule tables, from beginning
     /// of ChainedSequenceRuleSet table
@@ -394,27 +372,27 @@ ChainedSequenceRuleSet<'a> {
 }
 
 /// Part of [ChainedSequenceContextFormat1]
-ChainedSequenceRule<'a> {
+table ChainedSequenceRule {
     /// Number of glyphs in the backtrack sequence
-    #[compute_count(backtrack_sequence)]
+    #[compile(array_len($backtrack_sequence))]
     backtrack_glyph_count: BigEndian<u16>,
     /// Array of backtrack glyph IDs
     #[count(backtrack_glyph_count)]
     backtrack_sequence: [BigEndian<u16>],
     /// Number of glyphs in the input sequence
-    #[compute(plus_one(self.input_sequence.len()))]
+    #[compile(plus_one($input_sequence.len()))]
     input_glyph_count: BigEndian<u16>,
     /// Array of input glyph IDs—start with second glyph
-    #[count_with(minus_one, input_glyph_count)]
+    #[count_expr(minus_one($input_glyph_count))]
     input_sequence: [BigEndian<u16>],
     /// Number of glyphs in the lookahead sequence
-    #[compute_count(lookahead_sequence)]
+    #[compile(array_len($lookahead_sequence))]
     lookahead_glyph_count: BigEndian<u16>,
     /// Array of lookahead glyph IDs
     #[count(lookahead_glyph_count)]
     lookahead_sequence: [BigEndian<u16>],
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Array of SequenceLookupRecords
     #[count(seq_lookup_count)]
@@ -422,10 +400,9 @@ ChainedSequenceRule<'a> {
 }
 
 /// [Chained Sequence Context Format 2](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#chained-sequence-context-format-2-class-based-glyph-contexts)
-#[offset_host]
-ChainedSequenceContextFormat2<'a> {
+table ChainedSequenceContextFormat2 {
     /// Format identifier: format = 2
-    #[compute(2)]
+    #[format = 2]
     format: BigEndian<u16>,
     /// Offset to Coverage table, from beginning of
     /// ChainedSequenceContextFormat2 table
@@ -440,7 +417,7 @@ ChainedSequenceContextFormat2<'a> {
     /// from beginning of ChainedSequenceContextFormat2 table
     lookahead_class_def_offset: BigEndian<Offset16<ClassDef>>,
     /// Number of ChainedClassSequenceRuleSet tables
-    #[compute_count(chained_class_seq_rule_set_offsets)]
+    #[compile(array_len($chained_class_seq_rule_set_offsets))]
     chained_class_seq_rule_set_count: BigEndian<u16>,
     /// Array of offsets to ChainedClassSequenceRuleSet tables, from
     /// beginning of ChainedSequenceContextFormat2 table (may be NULL)
@@ -450,10 +427,9 @@ ChainedSequenceContextFormat2<'a> {
 }
 
 /// Part of [ChainedSequenceContextFormat2]
-#[offset_host]
-ChainedClassSequenceRuleSet<'a> {
+table ChainedClassSequenceRuleSet {
     /// Number of ChainedClassSequenceRule tables
-    #[compute_count(chained_class_seq_rule_offsets)]
+    #[compile(array_len($chained_class_seq_rule_offsets))]
     chained_class_seq_rule_count: BigEndian<u16>,
     /// Array of offsets to ChainedClassSequenceRule tables, from
     /// beginning of ChainedClassSequenceRuleSet
@@ -462,28 +438,28 @@ ChainedClassSequenceRuleSet<'a> {
 }
 
 /// Part of [ChainedSequenceContextFormat2]
-ChainedClassSequenceRule<'a> {
+table ChainedClassSequenceRule {
     /// Number of glyphs in the backtrack sequence
-    #[compute_count(backtrack_sequence)]
+    #[compile(array_len($backtrack_sequence))]
     backtrack_glyph_count: BigEndian<u16>,
     /// Array of backtrack-sequence classes
     #[count(backtrack_glyph_count)]
     backtrack_sequence: [BigEndian<u16>],
     /// Total number of glyphs in the input sequence
-    #[compute(plus_one(self.input_sequence.len()))]
+    #[compile(plus_one($input_sequence.len()))]
     input_glyph_count: BigEndian<u16>,
     /// Array of input sequence classes, beginning with the second
     /// glyph position
-    #[count_with(minus_one, input_glyph_count)]
+    #[count_expr(minus_one($input_glyph_count))]
     input_sequence: [BigEndian<u16>],
     /// Number of glyphs in the lookahead sequence
-    #[compute_count(lookahead_sequence)]
+    #[compile(array_len($lookahead_sequence))]
     lookahead_glyph_count: BigEndian<u16>,
     /// Array of lookahead-sequence classes
     #[count(lookahead_glyph_count)]
     lookahead_sequence: [BigEndian<u16>],
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Array of SequenceLookupRecords
     #[count(seq_lookup_count)]
@@ -491,63 +467,72 @@ ChainedClassSequenceRule<'a> {
 }
 
 /// [Chained Sequence Context Format 3](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#chained-sequence-context-format-3-coverage-based-glyph-contexts)
-#[offset_host]
-ChainedSequenceContextFormat3<'a> {
+table ChainedSequenceContextFormat3 {
     /// Format identifier: format = 3
-    #[compute(3)]
+    #[format = 3]
     format: BigEndian<u16>,
     /// Number of glyphs in the backtrack sequence
-    #[compute_count(backtrack_coverage_offsets)]
+    #[compile(array_len($backtrack_coverage_offsets))]
     backtrack_glyph_count: BigEndian<u16>,
     /// Array of offsets to coverage tables for the backtrack sequence
     #[count(backtrack_glyph_count)]
     backtrack_coverage_offsets: [BigEndian<Offset16<CoverageTable>>],
     /// Number of glyphs in the input sequence
-    #[compute_count(input_coverage_offsets)]
+    #[compile(array_len($input_coverage_offsets))]
     input_glyph_count: BigEndian<u16>,
     /// Array of offsets to coverage tables for the input sequence
     #[count(input_glyph_count)]
     input_coverage_offsets: [BigEndian<Offset16<CoverageTable>>],
     /// Number of glyphs in the lookahead sequence
-    #[compute_count(lookahead_coverage_offsets)]
+    #[compile(array_len($lookahead_coverage_offsets))]
     lookahead_glyph_count: BigEndian<u16>,
     /// Array of offsets to coverage tables for the lookahead sequence
     #[count(lookahead_glyph_count)]
     lookahead_coverage_offsets: [BigEndian<Offset16<CoverageTable>>],
     /// Number of SequenceLookupRecords
-    #[compute_count(seq_lookup_records)]
+    #[compile(array_len($seq_lookup_records))]
     seq_lookup_count: BigEndian<u16>,
     /// Array of SequenceLookupRecords
     #[count(seq_lookup_count)]
     seq_lookup_records: [SequenceLookupRecord],
 }
 
-#[format(u16)]
-#[offset_host]
-enum ChainedSequenceContext<'a> {
-    #[version(1)]
-    Format1(ChainedSequenceContextFormat1<'a>),
-    #[version(2)]
-    Format2(ChainedSequenceContextFormat2<'a>),
-    #[version(3)]
-    Format3(ChainedSequenceContextFormat3<'a>),
+format u16 ChainedSequenceContext {
+    Format1(ChainedSequenceContextFormat1),
+    Format2(ChainedSequenceContextFormat2),
+    Format3(ChainedSequenceContextFormat3),
+}
+
+/// [Device](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#device-and-variationindex-tables)
+/// delta formats
+enum u16 DeltaFormat {
+    /// Signed 2-bit value, 8 values per uint16
+    Local2BitDeltas = 0x0001,
+    /// Signed 4-bit value, 4 values per uint16
+    Local4BitDeltas = 0x0002,
+    /// Signed 8-bit value, 2 values per uint16
+    Local8BitDeltas = 0x0003,
+    /// VariationIndex table, contains a delta-set index pair.
+    VariationIndex = 0x8000,
+//0x7FFC	Reserved	For future use — set to 0
 }
 
 /// [Device Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#device-and-variationindex-tables)
-Device<'a> {
+table Device {
     /// Smallest size to correct, in ppem
     start_size: BigEndian<u16>,
     /// Largest size to correct, in ppem
     end_size: BigEndian<u16>,
     /// Format of deltaValue array data: 0x0001, 0x0002, or 0x0003
-    delta_format: BigEndian<u16>,
+    #[to_owned(convert_delta_format(obj.delta_format()))]
+    delta_format: BigEndian<DeltaFormat>,
     /// Array of compressed data
-    #[count_all]
+    #[count_expr(delta_value_count($start_size, $end_size, $delta_format))]
     delta_value: [BigEndian<u16>],
 }
 
 /// Variation index table
-VariationIndex {
+table VariationIndex {
     /// A delta-set outer index — used to select an item variation
     /// data subtable within the item variation store.
     delta_set_outer_index: BigEndian<u16>,
@@ -559,16 +544,11 @@ VariationIndex {
 }
 
 /// [FeatureVariations Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#featurevariations-table)
-#[offset_host]
-FeatureVariations<'a> {
-    /// Major version of the FeatureVariations table — set to 1.
-    #[compute(1)]
-    major_version: BigEndian<u16>,
-    /// Minor version of the FeatureVariations table — set to 0.
-    #[compute(0)]
-    minor_version: BigEndian<u16>,
+table FeatureVariations {
+    #[compile(MajorMinor::VERSION_1_0)]
+    version: BigEndian<MajorMinor>,
     /// Number of feature variation records.
-    #[compute_count(feature_variation_records)]
+    #[compile(array_len($feature_variation_records))]
     feature_variation_record_count: BigEndian<u32>,
     /// Array of feature variation records.
     #[count(feature_variation_record_count)]
@@ -576,7 +556,7 @@ FeatureVariations<'a> {
 }
 
 /// Part of [FeatureVariations]
-FeatureVariationRecord {
+record FeatureVariationRecord {
     /// Offset to a condition set table, from beginning of
     /// FeatureVariations table.
     condition_set_offset: BigEndian<Offset32<ConditionSet>>,
@@ -586,10 +566,9 @@ FeatureVariationRecord {
 }
 
 /// [ConditionSet Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#conditionset-table)
-#[offset_host]
-ConditionSet<'a> {
+table ConditionSet {
     /// Number of conditions for this condition set.
-    #[compute_count(condition_offsets)]
+    #[compile(array_len($condition_offsets))]
     condition_count: BigEndian<u16>,
     /// Array of offsets to condition tables, from beginning of the
     /// ConditionSet table.
@@ -604,9 +583,9 @@ ConditionSet<'a> {
 //}
 
 /// [Condition Table Format 1](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#condition-table-format-1-font-variation-axis-range): Font Variation Axis Range
-ConditionFormat1 {
+table ConditionFormat1 {
     /// Format, = 1
-    #[compute(1)]
+    #[format = 1]
     format: BigEndian<u16>,
     /// Index (zero-based) for the variation axis within the 'fvar'
     /// table.
@@ -620,36 +599,29 @@ ConditionFormat1 {
 }
 
 /// [FeatureTableSubstitution Table](https://docs.microsoft.com/en-us/typography/opentype/spec/chapter2#featuretablesubstitution-table)
-#[offset_host]
-FeatureTableSubstitution<'a> {
-    /// Major version of the feature table substitution table — set
-    /// to 1
-    #[compute(1)]
-    major_version: BigEndian<u16>,
-    /// Minor version of the feature table substitution table — set
-    /// to 0.
-    #[compute(0)]
-    minor_version: BigEndian<u16>,
+table FeatureTableSubstitution {
+    /// Major & minor version of the table: (1, 0)
+    #[compile(MajorMinor::VERSION_1_0)]
+    version: BigEndian<MajorMinor>,
     /// Number of feature table substitution records.
-    #[compute_count(substitutions)]
+    #[compile(array_len($substitutions))]
     substitution_count: BigEndian<u16>,
     /// Array of feature table substitution records.
     #[count(substitution_count)]
-    #[to_owned(self.substitutions_to_owned())]
     substitutions: [FeatureTableSubstitutionRecord],
 }
 
 /// Used in [FeatureTableSubstitution]
-#[skip_to_owned]
-FeatureTableSubstitutionRecord {
+record FeatureTableSubstitutionRecord {
     /// The feature table index to match.
     feature_index: BigEndian<u16>,
     /// Offset to an alternate feature table, from start of the
     /// FeatureTableSubstitution table.
+    #[skip_offset_getter] // custom impl, we need to pass a fake tag
     alternate_feature_offset: BigEndian<Offset32<Feature>>,
 }
 
-SizeParams {
+table SizeParams {
     /// The first value represents the design size in 720/inch units (decipoints).
     ///
     /// The design size entry must be non-zero. When there is a design size but
@@ -682,8 +654,9 @@ SizeParams {
     range_end: BigEndian<u16>,
 }
 
-StylisticSetParams {
-    #[compute(0)]
+table StylisticSetParams {
+    //#[version]
+    #[compile(0)]
     version: BigEndian<u16>,
     /// The 'name' table name ID that specifies a string (or strings, for
     /// multiple languages) for a user-interface label for this feature.
@@ -698,9 +671,9 @@ StylisticSetParams {
 }
 
 /// featureParams for ['cv01'-'cv99'](https://docs.microsoft.com/en-us/typography/opentype/spec/features_ae#cv01-cv99)
-CharacterVariantParams<'a> {
+table CharacterVariantParams {
     /// Format number is set to 0.
-    #[compute(0)]
+    #[format = 1]
     format: BigEndian<u16>,
     /// The 'name' table name ID that specifies a string (or strings,
     /// for multiple languages) for a user-interface label for this
@@ -721,7 +694,7 @@ CharacterVariantParams<'a> {
     first_param_ui_label_name_id: BigEndian<u16>,
     /// The count of characters for which this feature provides glyph
     /// variants. (May be zero.)
-    #[compute_count(character)]
+    #[compile(array_len($character))]
     char_count: BigEndian<u16>,
     /// The Unicode Scalar Value of the characters for which this
     /// feature provides glyph variants.
