--- resources/parse2_plan.toml
@@ -1,28 +1,0 @@
-# This is a 'codegen plan': it describes a series of steps to perform
-# in order to regenerate all of this project's generated code.
-
-# a list of directories that will be cleaned
-clean = ["font-tables/generated", "compile-tables/generated"]
-
-# TODO: support a 'copy' step
-
-# a mapping of files to generate, and their sources
-[[generate]]
-mode = "parse2"
-source = "resources/codegen_inputs/gpos2.rs"
-target = "font-tables/generated/gpos.rs"
-
-[[generate]]
-mode = "parse2"
-source = "resources/codegen_inputs/layout2.rs"
-target = "font-tables/generated/layout.rs"
-
-[[generate]]
-mode = "compile2"
-source = "resources/codegen_inputs/gpos2.rs"
-target = "compile-tables/generated/gpos.rs"
-
-[[generate]]
-mode = "compile2"
-source = "resources/codegen_inputs/layout2.rs"
-target = "compile-tables/generated/layout.rs"
